+++ packages/moonstone/TooltipDecorator/TooltipDecorator.js
@@ -200,12 +200,12 @@ const TooltipDecorator = hoc((config, Wrapped) => {
 		}
 
 		adjustAnchor (arrowAnchor, tooltipDirection, overflow) {
-			if (this.context.rtl && (tooltipDirection === 'above' || tooltipDirection === 'below')) {
-				arrowAnchor = arrowAnchor === 'left' ? 'right' : 'left';
-			}
-
-			// Flip sideways for 'above' and 'below' if it overflows to the sides
 			if (tooltipDirection === 'above' || tooltipDirection === 'below') {
+				if (this.context.rtl && arrowAnchor !== 'center') {
+					arrowAnchor = arrowAnchor === 'left' ? 'right' : 'left';
+				}
+
+				// Flip sideways if it overflows to the sides
 				if (overflow.isOverRight) {
 					arrowAnchor = 'left';
 				} else if (overflow.isOverLeft) {