+++ packages/moonstone/CHANGELOG.md
@@ -7,12 +7,15 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Changed
 
 - `moonstone/Picker` component's buttons to reverse their operation such that 'up' selects the previous item and 'down' the next
+### Added
+
+- `moonstone/MarqueeDecorator` property `invalidateProps` that specifies which props cause the marquee distance to be invalidated.
 
 ## [1.0.0-beta.1] - 2016-12-30
 
 ### Added
 
-- `Moonstone/VideoPlayer` and `moonstone/TooltipDecorator` components and samples
+- `moonstone/VideoPlayer` and `moonstone/TooltipDecorator` components and samples
 - `moonstone/Panels.Panels` property `onBack` to support `ui/Cancelable`
 - `moonstone/VirtualFlexList` Work-In-Progress component to support variably sized rows or columns
 - `moonstone/ExpandableItem` properties `autoClose` and `lockBottom`